making export and env more bash_like
-> export is able to create envars without value
-> env should not display envars without value
-> export without arguments should display even variables without value
with "declare -x" as its prefix

	modified:   Makefile
	modified:   include/minishell.h
	modified:   src/builtins/env_builtin.c
	modified:   src/builtins/export_builtin.c
	modified:   src/builtins/utils_builtins.c
	deleted:    src/utils/get_env_content.c
