+++ include/minishell.h
@@ -166,4 +166,5 @@ int						is_being_initialized(char *new_env);
 int						already_exists(t_databus *data, char *env);
 t_fn_built_exec			**get_built_func(void);
 int						names_are_equal(char *s1, char *s2);
+void					get_env_name(t_databus *data, char *env, char *name);
 #endif // !
